Report management UI skeleon.
A skeleton of a Next.JS app for managing reports.  The app loads up wit
a report listing page, and includes a page to create a new report, and a
page to view the report details.

As of now, the UI uses mocked up data, and DOES NOT make any calls to the reporting
service API.

NO_GLINT=true

Change-Id: I2ace22be39dc67b7b41166614f456ec0fddd9c75